split out component
